requirements.txt
bitcoin==1.1.42
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
firebase-admin==6.4.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
future==0.18.3
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.0.7
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycoin==0.92.20230326
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
requests==2.31.0
rsa==4.9
setuptools==69.0.3
transaction==4.0
uritemplate==4.1.1
urllib3==2.2.0
Werkzeug==3.0.1
zope.interface==6.1
